requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
tabula-py==2.9.0
reportlab==4.0.7
PyPDF2==3.0.1
xlrd==2.0.1
jinja2==3.1.2
aiofiles==23.2.1
